requirements.txt
sqlalchemy>=2.0
streamlit>=1.0
psycopg2-binary>=2.9
matplotlib>=3.8
python-dotenv>=1.0
pandas>=2.0
pytest>=7.0
pytest-cov>=4.0
